refactor: split
