+++ runner/src/lib.rs
@@ -1,6 +1,7 @@
 use env::{PERMISSION_ID, RUNNING_PATH, RunnerOption, SH_CMD};
-use lang::{LangExt, Runners, runner::RunCommand};
+use lang::LangExt;
 use nsjail::NsJailBuilder;
+use runner::{RunCommand, Runners};
 use std::{
     io::Write,
     path::{Path, PathBuf},
@@ -16,6 +17,7 @@ use runner_schema::{
 
 pub mod lang;
 pub mod nsjail;
+pub mod runner;
 pub mod time;
 
 pub mod env;
